# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WEREAL REOS API",
    "version": "1.0.0",
    "description": "REST API contract for WEREAL Real Estate Operating System — Phase 1 MVP.\n\nGenerated from `Thiet-ke-API.md` (WEREAL-API-2026-v1.0).\nBaseline: WEREAL-BL-2026-002.\n\n71 endpoints across Identity, Golden Record, Listing, Search, CRM, Booking, Payment, Ledger, AI, Analytics, Audit, and Stream modules.",
    "contact": {
      "name": "WEREAL Platform Team",
      "email": "platform@wereal.vn"
    },
    "license": {
      "name": "Proprietary"
    },
    "x-document-id": "WEREAL-API-2026-v1.0"
  },
...
  "servers": [
    {
      "url": "https://api.wereal.vn/api/v1",
      "description": "Production"
    },
    {
      "url": "https://api.staging.wereal.vn/api/v1",
      "description": "Staging"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/leads": {
      "get": {
        "tags": [
          "CRM"
        ],
        "operationId": "get_leads",
        "summary": "API-041: GET /leads",
        "description": "**API-041** | Module: CRM | FR: FR-CRM-01,03 | Auth: Agent / Agency Admin",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/XRequestId"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "data": [
                    {
                      "id": "ld_01",
                      "attributes": {
                        "fullName": "Thu Trang",
                        "phone": "+849***",
                        "score": 85,
                        "tier": "HOT",
                        "source": "PUBLIC_FORM"
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "x-wereal-api-id": "API-041",
        "x-wereal-fr-link": "FR-CRM-01,03"
      },
      "post": {
        "tags": [
          "CRM"
        ],
        "operationId": "post_leads",
        "summary": "API-042: POST /leads",
        "description": "**API-042** | Module: CRM | FR: FR-CRM-01, BR-15, NFR-C03 | Auth: Public (lead form) / Agent",
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestId"
          },
          {
            "$ref": "#/components/parameters/XIdempotencyKey"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "data": {
                    "id": "ld_01",
                    "attributes": {
                      "status": "NEW",
                      "score": null,
                      "routingStatus": "PENDING"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetails"
          }
        },
        "x-wereal-api-id": "API-042",
        "x-wereal-fr-link": "FR-CRM-01, BR-15, NFR-C03",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericObject"
              },
              "example": {
                "fullName": "Nguyễn Thu Trang",
                "phone": "+84901234567",
                "email": "trang@gmail.com",
                "listingId": "ls_01",
                "unitId": "un_01",
                "message": "Muốn xem căn 3PN",
                "consent": {
                  "marketing": true,
                  "privacyPolicyVersion": "2026-07-01"
                },
                "utm": {
                  "source": "google",
                  "campaign": "q7_launch"
                }
              }
            }
          }
        }
      }
    },
    "/leads/{leadId}": {
      "get": {
        "tags": [
          "CRM"
        ],
        "operationId": "get_leads_leadId",
        "summary": "API-043: GET /leads/{leadId}",
        "description": "**API-043** | Module: CRM | FR: FR-CRM-02 | Auth: Agent (assigned)",
        "parameters": [
          {
            "name": "leadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Resource identifier (leadId)"
          },
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/XRequestId"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "data": {
                    "id": "ld_01",
                    "relationships": {
                      "unit": {
                        "data": {
                          "id": "un_01"
                        }
                      },
                      "activities": {
                        "meta": {
                          "count": 5
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetails"
          }
        },
        "x-wereal-api-id": "API-043",
        "x-wereal-fr-link": "FR-CRM-02",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "CRM"
        ],
        "operationId": "patch_leads_leadId",
        "summary": "API-044: PATCH /leads/{leadId}",
        "description": "**API-044** | Module: CRM | FR: FR-CRM-05 | Auth: Agent",
        "parameters": [
          {
            "name": "leadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Resource identifier (leadId)"
          },
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "data": {
                    "id": "ld_01",
                    "attributes": {
                      "stage": "QUALIFIED"
                    }
                  }
                }
              }
            }
          }
        },
        "x-wereal-api-id": "API-044",
        "x-wereal-fr-link": "FR-CRM-05",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericObject"
              },
              "example": {
                "stage": "QUALIFIED",
                "notes": "Khách có nhu cầu mua trong T8",
                "tags": [
                  "VIP"
                ]
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "CRM"
        ],
        "operationId": "delete_leads_leadId",
        "summary": "API-045: DELETE /leads/{leadId}",
        "description": "**API-045** | Module: CRM | FR: FR-CRM-01 | Auth: Agency Admin",
        "parameters": [
          {
            "name": "leadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Resource identifier (leadId)"
          },
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "x-wereal-api-id": "API-045",
        "x-wereal-fr-link": "FR-CRM-01",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/tenants": {
      "get": {
        "tags": [
          "Identity"
        ],
        "operationId": "get_tenants",
        "summary": "API-004: GET /tenants",
        "description": "**API-004** | Module: Identity | FR: FR-ID-01, US-ID-01,02 | Auth: Platform Admin",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/XRequestId"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "data": [
                    {
                      "id": "ten_dev_01",
                      "type": "tenant",
                      "attributes": {
                        "name": "Vinhomes Pilot",
                        "type": "DEVELOPER",
                        "status": "ACTIVE"
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ProblemDetails"
          }
        },
        "x-wereal-api-id": "API-004",
        "x-wereal-fr-link": "FR-ID-01, US-ID-01,02",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Identity"
        ],
        "operationId": "post_tenants",
        "summary": "API-005: POST /tenants",
        "description": "**API-005** | Module: Identity | FR: FR-ID-01, UC-ID-01 | Auth: Platform Admin",
        "parameters": [
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/XRequestId"
          },
          {
            "$ref": "#/components/parameters/XIdempotencyKey"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "data": {
                    "id": "ten_agy_01",
                    "type": "tenant",
                    "attributes": {
                      "name": "Agency Pilot ABC",
                      "status": "PENDING_SETUP"
                    }
                  }
                }
              }
            }
          },
          "409": {
            "$ref": "#/components/responses/ProblemDetails"
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetails"
          }
        },
        "x-wereal-api-id": "API-005",
        "x-wereal-fr-link": "FR-ID-01, UC-ID-01",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericObject"
              },
              "example": {
                "name": "Agency Pilot ABC",
                "type": "AGENCY",
                "slug": "agency-abc",
                "adminEmail": "admin@abc.vn",
                "settings": {
                  "timezone": "Asia/Ho_Chi_Minh",
                  "locale": "vi-VN"
                }
              }
            }
          }
        }
      }
    },
    "/tenants/{tenantId}": {
      "get": {
        "tags": [
          "Identity"
        ],
        "operationId": "get_tenants_tenantId",
        "summary": "API-006: GET /tenants/{tenantId}",
        "description": "**API-006** | Module: Identity | FR: FR-ID-01 | Auth: Platform Admin / Tenant Admin",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Resource identifier (tenantId)"
          },
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/XRequestId"
          },
          {
            "$ref": "#/components/parameters/Cursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "data": {
                    "id": "ten_dev_01",
                    "attributes": {
                      "name": "Dev Pilot",
                      "type": "DEVELOPER",
                      "status": "ACTIVE"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/ProblemDetails"
          }
        },
        "x-wereal-api-id": "API-006",
        "x-wereal-fr-link": "FR-ID-01",
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Identity"
        ],
        "operationId": "patch_tenants_tenantId",
        "summary": "API-007: PATCH /tenants/{tenantId}",
        "description": "**API-007** | Module: Identity | FR: FR-ID-01 | Auth: Platform Admin",
        "parameters": [
          {
            "name": "tenantId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Resource identifier (tenantId)"
          },
          {
            "$ref": "#/components/parameters/XTenantId"
          },
          {
            "$ref": "#/components/parameters/XRequestId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "data": {
                    "id": "ten_dev_01",
                    "attributes": {
                      "status": "ACTIVE"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ProblemDetails"
          }
        },
        "x-wereal-api-id": "API-007",
        "x-wereal-fr-link": "FR-ID-01",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenericObject"
              },
              "example": {
                "status": "ACTIVE",
                "settings": {
                  "bookingExpiryHours": 48
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token (TTL 15 minutes). Requires X-Tenant-Id header."
      },
      "RefreshToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Refresh-Token",
        "description": "Refresh token (also accepted in request body or httpOnly cookie)."
      },
      "WebhookHmac": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Signature",
        "description": "HMAC-SHA256 signature + gateway IP allowlist."
      }
    },
    "parameters": {
      "XTenantId": {
        "name": "X-Tenant-Id",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Tenant context for RLS isolation (FR-ID-03)."
      },
      "XRequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Trace correlation ID (NFR-O01)."
      },
      "XIdempotencyKey": {
        "name": "X-Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Idempotency key for safe retries (TTL 24h)."
      },
      "Cursor": {
        "name": "cursor",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Cursor for pagination."
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        },
        "description": "Page size (max 100)."
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Resource"
              },
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Resource"
                }
              },
              {
                "type": "object"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        }
      },
      "Resource": {
        "type": "object",
        "required": [
          "id",
          "type"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Prefixed ULID (un_, ls_, bk_, ld_)"
          },
          "type": {
            "type": "string"
          },
          "attributes": {
            "type": "object",
            "additionalProperties": true
          },
          "relationships": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "ResponseMeta": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "page": {
            "$ref": "#/components/schemas/PageMeta"
          }
        }
      },
      "PageMeta": {
        "type": "object",
        "properties": {
          "cursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "hasMore": {
            "type": "boolean"
          },
          "limit": {
            "type": "integer"
          },
          "number": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          },
          "totalElements": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 Problem Details",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "traceId": {
            "type": "string",
            "format": "uuid"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "GenericObject": {
        "type": "object",
        "additionalProperties": true
      }
    },
    "responses": {
      "ProblemDetails": {
        "description": "Error response (RFC 7807)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      }
    }
  },
...
}